# server/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User Management API",
        "version": "1.0.0",
        "description": "API for user authentication and management, including account verification, password resets, and user information updates."
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1"
        }
    ],
    "components": {
        "schemas": {
...
            "Course": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "thumbnail": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "level": {
                        "type": "string"
                    },
                    "courseData": {
                        "type": "object",
                        "additionalProperties": true
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "CourseInput": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "thumbnail": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "level": {
                        "type": "string"
                    },
                    "courseData": {
                        "type": "object",
                        "additionalProperties": true
                    }
                },
                "required": [
                    "name",
                    "description",
                    "price"
                ]
            }
        }
    },
    "paths": {
...
        "/courses/getVdoCipherOTP": {
            "post": {
                "summary": "Generate video OTP",
                "operationId": "generateVideoUrl",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "videoId": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "videoId"
                                ]
                            }
                        }
                    },
                    "responses": {
                        "200": {
                            "description": "OTP generated successfully",
                            "content": {
                                "application/json": {
                                    "schema": {
                                        "type": "object",
                                        "additionalProperties": true
                                    }
                                }
                            }
                        },
                        "400": {
                            "description": "Error in generating OTP"
                        }
                    }
                }
            }
        },
        "/courses/get-course/{id}": {
            "get": {
                "summary": "Get single course by ID",
                "operationId": "getSingleCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "course": {
                                            "$ref": "#/components/schemas/Course"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found"
                    }
                }
            }
        },
        "/courses/get-courses": {
            "get": {
                "summary": "Get all courses",
                "operationId": "getAllCourses",
                "responses": {
                    "200": {
                        "description": "List of courses retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "count": {
                                            "type": "integer"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "courses": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Course"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/courses/get-course-content/{id}": {
            "get": {
                "summary": "Get course content by user",
                "operationId": "getCourseByUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course content retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "content": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not eligible to access this course"
                    }
                }
            }
        },
        "/courses/add-question": {
            "put": {
                "summary": "Add a question to a course",
                "operationId": "addQuestion",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "question": {
                                        "type": "string"
                                    },
                                    "courseId": {
                                        "type": "string"
                                    },
                                    "contentId": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "question",
                                    "courseId",
                                    "contentId"
                                ]
                            }
                        }
                    },
                    "responses": {
                        "200": {
                            "description": "Question added successfully",
                            "content": {
                                "application/json": {
                                    "schema": {
                                        "type": "object",
                                        "properties": {
                                            "success": {
                                                "type": "boolean"
                                            },
                                            "course": {
                                                "$ref": "#/components/schemas/Course"
                                            }
                                        }
                                    }
                                }
                            }
                        },
                        "400": {
                            "description": "Invalid content id"
                        },
                        "401": {
                            "description": "User not authenticated"
                        },
                        "404": {
                            "description": "Course not found"
                        },
                        "500": {
                            "description": "Server error"
                        }
                    }
                }
            }
        },
        "/courses/add-answer": {
            "put": {
                "summary": "Add an answer to a question in a course",
                "operationId": "addAnswer",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "answer": {
                                        "type": "string"
                                    },
                                    "courseId": {
                                        "type": "string"
                                    },
                                    "contentId": {
                                        "type": "string"
                                    },
                                    "questionId": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "answer",
                                    "courseId",
                                    "contentId",
                                    "questionId"
                                ]
                            }
                        }
                    },
                    "responses": {
                        "200": {
                            "description": "Answer added successfully",
                            "content": {
                                "application/json": {
                                    "schema": {
                                        "type": "object",
                                        "properties": {
                                            "success": {
                                                "type": "boolean"
                                            },
                                            "course": {
                                                "$ref": "#/components/schemas/Course"
                                            }
                                        }
                                    }
                                }
                            }
                        },
                        "400": {
                            "description": "Invalid content id or question id"
                        },
                        "401": {
                            "description": "User not authenticated"
                        },
                        "404": {
                            "description": "Course or question not found"
                        },
                        "500": {
                            "description": "Server error"
                        }
                    }
                }
            }
        },
        "/courses/create-course": {
            "post": {
                "summary": "Upload a new course",
                "operationId": "uploadCourse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CourseInput"
                            }
                        }
                    },
                    "responses": {
                        "201": {
                            "description": "Course created successfully",
                            "content": {
                                "application/json": {
                                    "schema": {
                                        "type": "object",
                                        "properties": {
                                            "success": {
                                                "type": "boolean"
                                            },
                                            "course": {
                                                "$ref": "#/components/schemas/Course"
                                            }
                                        }
                                    }
                                }
                            }
                        },
                        "400": {
                            "description": "Error in creating course"
                        },
                        "401": {
                            "description": "User not authenticated"
                        },
                        "500": {
                            "description": "Server error"
                        }
                    }
                }
            }
        },
        "/courses/edit-course/{id}": {
            "put": {
                "summary": "Edit a course by ID",
                "operationId": "editCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CourseInput"
                            }
                        }
                    },
                    "responses": {
                        "200": {
                            "description": "Course edited successfully",
                            "content": {
                                "application/json": {
                                    "schema": {
                                        "type": "object",
                                        "properties": {
                                            "success": {
                                                "type": "boolean"
                                            },
                                            "course": {
                                                "$ref": "#/components/schemas/Course"
                                            }
                                        }
                                    }
                                }
                            }
                        },
                        "404": {
                            "description": "Course not found"
                        },
                        "400": {
                            "description": "Invalid input"
                        },
                        "401": {
                            "description": "User not authenticated"
                        },
                        "500": {
                            "description": "Server error"
                        }
                    }
                }
            }
        },
        "/courses/delete-course/{id}": {
            "delete": {
                "summary": "Delete a course by ID",
                "operationId": "deleteCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Course deleted successfully"
                    },
                    "404": {
                        "description": "Course not found"
                    },
                    "401": {
                        "description": "User not authenticated"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }
    }
}